Renamed package to match the Google Play listing name
- Renamed package to match the Google Play listing name
- Make FAB add button invisible when add dialog is open
