+++ horizon-lite/app/build.gradle
@@ -7,7 +7,7 @@ android {
     buildToolsVersion "30.0.3"
 
     defaultConfig {
-        applicationId "com.innerCat.sunrise"
+        applicationId "com.innerCat.sunset"
         minSdkVersion 26
         targetSdkVersion 30
         versionCode 1
