+++ horizon-lite/app/src/androidTest/java/com/innerCat/sunset/ExampleInstrumentedTest.java
@@ -1,4 +1,4 @@
-package com.innerCat.sunrise;
+package com.innerCat.sunset;
 
 import android.content.Context;
