+++ horizon-lite/app/src/main/java/com/innerCat/sunset/room/Converters.java
@@ -1,4 +1,4 @@
-package com.innerCat.sunrise.room;
+package com.innerCat.sunset.room;
 
 import android.content.res.Resources;
 import android.util.TypedValue;
