+++ horizon-lite/app/src/main/java/com/innerCat/sunset/room/TaskDatabase.java
@@ -1,10 +1,10 @@
-package com.innerCat.sunrise.room;
+package com.innerCat.sunset.room;
 
 import androidx.room.Database;
 import androidx.room.RoomDatabase;
 import androidx.room.TypeConverters;
 
-import com.innerCat.sunrise.Task;
+import com.innerCat.sunset.Task;
 
 @Database(entities = { Task.class}, version = 2)
 @TypeConverters({ Converters.class})
