+++ horizon-lite/app/src/main/java/com/innerCat/sunset/widgets/DataProvider.java
@@ -1,24 +1,18 @@
-package com.innerCat.sunrise.widgets;
+package com.innerCat.sunset.widgets;
 
-import android.app.PendingIntent;
 import android.content.Context;
 import android.content.Intent;
-import android.os.Handler;
-import android.os.Looper;
 import android.widget.RemoteViews;
 import android.widget.RemoteViewsService;
 
 import androidx.room.Room;
 
-import com.innerCat.sunrise.R;
-import com.innerCat.sunrise.Task;
-import com.innerCat.sunrise.activities.MainActivity;
-import com.innerCat.sunrise.room.TaskDatabase;
+import com.innerCat.sunset.R;
+import com.innerCat.sunset.Task;
+import com.innerCat.sunset.room.TaskDatabase;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.concurrent.ExecutorService;
-import java.util.concurrent.Executors;
 
 
 public class DataProvider implements RemoteViewsService.RemoteViewsFactory {
