+++ horizon-lite/app/src/main/res/layout/activity_main.xml
@@ -5,7 +5,8 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         tools:context=".activities.MainActivity"
-        android:backgroundTintMode="add">
+        android:backgroundTintMode="add"
+        android:animateLayoutChanges="true">
 
     <com.google.android.material.floatingactionbutton.FloatingActionButton
             android:layout_width="wrap_content"
