+++ horizon-lite/app/src/main/res/values-night/themes.xml
@@ -32,7 +32,7 @@
         <item name="android:textColor">#FFFFFF</item>
     </style>
 
-    <style name="ThemeOverlay.Sunrise.AppWidgetContainer" parent="">
+    <style name="ThemeOverlay.Sunset.AppWidgetContainer" parent="">
         <item name="appWidgetBackgroundColor">@color/light_blue_900</item>
         <item name="appWidgetTextColor">@color/light_blue_200</item>
     </style>
