+++ horizon-lite/app/src/main/res/values/strings.xml
@@ -1,6 +1,6 @@
 <resources>
     <!-- <string name="app_name">horizon-lite</string> -->
-    <string name="app_name">Sunrise</string>
+    <string name="app_name">Sunset</string>
     <string name="streak">streak</string>
     <string name="last_updated">lastUpdated</string>
     <string name="increase_streak">increaseStreak</string>
