+++ horizon-lite/app/src/main/res/values/themes.xml
@@ -32,7 +32,7 @@
         <item name="android:textColor">#000000</item>
     </style>
 
-    <style name="ThemeOverlay.Sunrise.AppWidgetContainer" parent="">
+    <style name="ThemeOverlay.Sunset.AppWidgetContainer" parent="">
         <item name="appWidgetBackgroundColor">@color/light_blue_600</item>
         <item name="appWidgetTextColor">@color/light_blue_50</item>
     </style>
