+++ horizon-lite/settings.gradle
@@ -1,2 +1,2 @@
 include ':app'
-rootProject.name = "sunrise"
+rootProject.name = "sunset"
